requirements.txt
fastapi
uvicorn
pymupdf
pytesseract
pdf2image
python-docx
requests
python-multipart
dotenv
pytest
